requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
numpy>=1.26.0
scikit-learn>=1.3.2
joblib>=1.3.2
lightgbm>=4.1.0
pandas>=2.1.3
supabase>=2.0.0
python-dotenv>=1.0.0

